requirements.txt
Click
pytest
opencv-python
pdfrw
Pillow
pygetwindow
pyautogui
inject